twister: rename platform.py
Changed platform.py to twister_platform.py.
On some OSes, like Windows,
the platform name collides with built-in platform module
and causes errors with pickling used in Black Box tests.
