+++ scripts/pylib/twister/twisterlib/handlers.py
@@ -27,8 +27,8 @@ from queue import Empty, Queue
 import psutil
 from twisterlib.environment import ZEPHYR_BASE, strip_ansi_sequences
 from twisterlib.error import TwisterException
-from twisterlib.platform import Platform
 from twisterlib.statuses import TwisterStatus
+from twisterlib.twister_platform import Platform
 
 sys.path.insert(0, os.path.join(ZEPHYR_BASE, "scripts/pylib/build_helpers"))
 from domains import Domains
